# Documentazione/Progettazione/openapi.yaml
openapi: 3.0.3
info:
  title: Monopoly IoT API
  version: 1.0.0
  description: |
    API REST esposte dall'API Gateway del sistema Monopoly IoT.

    Il sistema è composto da 5 microservizi:
    - **api-gateway** (porta 8080) — punto di ingresso unico
    - **user-service** (porta 8081) — gestione utenti
    - **match-service** (porta 8082) — partite, tornei, motore di gioco
    - **board-service** (porta 8083) — sedi, tavoli, zone RFID
    - **iot-gateway** (porta 8085) — acquisizione eventi MQTT, buffer offline

    Tutte le richieste vengono inoltrate dal Gateway ai servizi interni.
servers:
  - url: http://localhost:8080
    description: API Gateway locale
...
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: Elenco utenti
      operationId: listUsers
      responses:
        "200":
          description: Lista utenti
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      tags: [Users]
      summary: Crea utente (registrazione)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
            example:
              username: "mario_rossi"
              email: "mario@example.com"
              passwordHash: "abc123"
              fullName: "Mario Rossi"
              active: true
              wantsTournament: false
      responses:
        "200":
          description: Utente creato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /api/venues:
    get:
      tags: [Board]
      summary: Elenco sedi
      operationId: listVenues
      responses:
        "200":
          description: Lista sedi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venue"
    post:
      tags: [Board]
      summary: Crea sede
      operationId: createVenue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Venue"
      responses:
        "200":
          description: Sede creata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
  /api/venues/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Board]
      summary: Dettaglio sede
      operationId: getVenueById
      responses:
        "200":
          description: Sede trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "404":
          description: Sede non trovata
    put:
      tags: [Board]
      summary: Aggiorna sede
      operationId: updateVenue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Venue"
      responses:
        "200":
          description: Sede aggiornata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
    delete:
      tags: [Board]
      summary: Elimina sede
      operationId: deleteVenue
      responses:
        "200":
          description: Sede eliminata
...
  /api/gametables:
    get:
      tags: [Board]
      summary: Elenco tavoli di gioco
      operationId: listGameTables
      responses:
        "200":
          description: Lista tavoli
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameTable"
    post:
      tags: [Board]
      summary: Crea tavolo di gioco
      description: |
        Crea un nuovo tavolo e clona automaticamente le ReadingZone
        del tavolo template (id=1), aggiornando i prefissi dei codici sensore.
      operationId: createGameTable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameTable"
            example:
              venueId: 1
              status: "ONLINE"
              displayName: "Tavolo Nexus-02"
              tableCode: "TBL002"
      responses:
        "200":
          description: Tavolo creato con zone RFID clonate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameTable"
  /api/gametables/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Board]
      summary: Dettaglio tavolo
      operationId: getGameTableById
      responses:
        "200":
          description: Tavolo trovato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameTable"
        "404":
          description: Tavolo non trovato
    put:
      tags: [Board]
      summary: Aggiorna tavolo
      operationId: updateGameTable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameTable"
      responses:
        "200":
          description: Tavolo aggiornato
    delete:
      tags: [Board]
      summary: Elimina tavolo
      operationId: deleteGameTable
      responses:
        "200":
          description: Tavolo eliminato
...
  /api/game-settings:
    get:
      tags: [Game Settings]
      summary: Legge le impostazioni di gioco attive
      description: Se non esistono impostazioni, ne crea una di default.
      operationId: getGameSettings
      responses:
        "200":
          description: Impostazioni correnti
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSettings"
    put:
      tags: [Game Settings]
      summary: Aggiorna impostazioni di gioco
      operationId: updateGameSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameSettings"
            example:
              initialBalance: 1500
              maxPlayers: 4
              sensorModeActive: true
      responses:
        "200":
          description: Impostazioni aggiornate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSettings"

  # ─────────────────────────────────────────────────────────────
  # IOT SYNC (→ iot-gateway :8085)
  # ─────────────────────────────────────────────────────────────
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Identificatore univoco della risorsa
  schemas:
    User:
      type: object
      description: Utente della piattaforma (giocatore, venue manager o admin)
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        username:
          type: string
          example: "mario_rossi"
        email:
          type: string
          format: email
          example: "mario@example.com"
        passwordHash:
          type: string
          description: Hash della password (non restituire in produzione)
        status:
          type: string
          example: "ONLINE"
        fullName:
          type: string
          example: "Mario Rossi"
        active:
          type: boolean
          default: true
        wantsTournament:
          type: boolean
          default: false
          description: true se il giocatore è disponibile per tornei
        bio:
          type: string
        gender:
          type: string
        age:
          type: integer
        avatar:
          type: string
          description: URL o base64 dell'immagine avatar
        adminNotice:
          type: string
          description: Comunicazione mostrata al prossimo accesso
    Venue:
      type: object
      description: Sede fisica dove si trovano i tavoli Monopoly IoT
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
          example: "Sala Giochi Centrale"
        location:
          type: string
        address:
          type: string
        city:
          type: string
          example: "Milano"
        type:
          type: string
          description: Tipo di sede (es. "Sala Giochi", "Bar", "Circolo")
        contactInfo:
          type: string
        ownerUserId:
          type: integer
          format: int64
          description: userId del Venue Manager responsabile della sede
...
    GameTable:
      type: object
      description: Tavolo fisico di gioco con sensori RFID
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        venueId:
          type: integer
          format: int64
        status:
          type: string
          enum: [ONLINE, OFFLINE, MAINTENANCE]
          example: "ONLINE"
        displayName:
          type: string
          example: "Tavolo Nexus-01"
        tableCode:
          type: string
          example: "TBL001"
        gameTypeId:
          type: integer
          format: int64
          default: 1
          description: Tipo di gioco (1 = Monopoly)
...
    GameSettings:
      type: object
      description: Impostazioni globali di gioco (singleton)
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        initialBalance:
          type: integer
          default: 1500
          description: Saldo iniziale in lire Monopoly per ogni giocatore
        maxPlayers:
          type: integer
          default: 4
          description: Numero massimo di giocatori per partita
        sensorModeActive:
          type: boolean
          default: true
          description: true = eventi fisici RFID; false = modalità manuale
...